requirements.txt
flask
gunicorn
mysql-connector-python
pytest
python-dotenv
selenium
